moving qt utilities into qtutils module
