+++ packages/pymodaq/src/pymodaq/control_modules/control_module_selector.py
@@ -1,4 +1,6 @@
 from qtpy import QtWidgets, QtCore
+
+import pymodaq_gui.qt_utils
 from pymodaq.control_modules.instruments import DET_TYPES
 
 REMOTE_ITEMS  = {'LECODirector', 'TCPServer'}
@@ -129,7 +131,7 @@ def add_category_layers(dimension_dict, remote_items=None, mock_items=None):
 
 
 if __name__ == '__main__':
-    from pymodaq_gui.utils.utils import mkQApp
+    from pymodaq_gui.qt_utils import mkQApp
 
     app = mkQApp('Selector')
 
@@ -145,6 +147,6 @@ if __name__ == '__main__':
     selector = ModuleSelector('Add', add_menu_entries)
     selector.add_widget.show()
 
-    app.exec()
+    pymodaq_gui.qt_utils.exec()
