+++ packages/pymodaq/src/pymodaq/extensions/adaptive/adaptive_optimization.py
@@ -1,3 +1,4 @@
+import pymodaq_gui.qt_utils
 from pymodaq_gui.messenger import messagebox
 from pymodaq_utils import utils
 from pymodaq_utils import config as config_mod
@@ -162,7 +163,7 @@ class AdaptiveOptimisation(GenericOptimization):
 
 
 def main():
-    from pymodaq_gui.utils.utils import mkQApp
+    from pymodaq_gui.qt_utils import mkQApp
     from pymodaq.utils.gui_utils.loader_utils import load_dashboard_with_preset
 
     app = mkQApp('Adaptive Optimiser')
@@ -170,7 +171,7 @@ def main():
 
     dashboard, extension, win = load_dashboard_with_preset(preset_file_name, 'AdaptiveScan')
 
-    app.exec()
+    pymodaq_gui.qt_utils.exec()
 
     return dashboard, extension, win
