+++ packages/pymodaq/src/pymodaq/extensions/console.py
@@ -8,6 +8,7 @@ Created the 25/10/2022
 from qtconsole.rich_jupyter_widget import RichJupyterWidget
 from qtconsole.inprocess import QtInProcessKernelManager
 
+import pymodaq_gui.qt_utils
 from pymodaq_utils import config as configmod
 from pymodaq_utils.utils import get_version
 
@@ -56,7 +57,7 @@ class QtConsole(RichJupyterWidget):
 
 
 def main():
-    from pymodaq_gui.utils.utils import mkQApp
+    from pymodaq_gui.qt_utils import mkQApp
 
     app = mkQApp('Console')
 
@@ -65,7 +66,7 @@ def main():
                      custom_banner=BANNER,
                      )
     prog.show()
-    app.exec()
+    pymodaq_gui.qt_utils.exec()
     return prog
