+++ packages/pymodaq/src/pymodaq/extensions/daq_scan.py
@@ -19,6 +19,7 @@ from qtpy import QtWidgets, QtCore, QtGui
 from qtpy.QtWidgets import QDialogButtonBox
 from qtpy.QtCore import QObject, Slot, QThread, Signal, QDateTime, QDate, QTime
 
+import pymodaq_gui.qt_utils
 from pymodaq_utils.logger import set_logger, get_module_name
 from pymodaq_utils.config import Config
 from pymodaq_utils import utils
@@ -1199,7 +1200,7 @@ class DAQScanAcquisition(QObject):
 
 
 def main():
-    from pymodaq_gui.utils.utils import mkQApp
+    from pymodaq_gui.qt_utils import mkQApp
     from pymodaq.utils.gui_utils.loader_utils import load_dashboard_with_preset
 
     app = mkQApp('DAQScan')
@@ -1207,7 +1208,7 @@ def main():
 
     dashboard, extension, win = load_dashboard_with_preset(preset_file_name, 'DAQScan')
 
-    app.exec()
+    pymodaq_gui.qt_utils.exec()
 
     return dashboard, extension, win
