+++ packages/pymodaq/src/pymodaq/extensions/pid/utils.py
@@ -137,7 +137,7 @@ def main(xmlfile):
     from pymodaq.utils.gui_utils.loader_utils import load_dashboard_with_preset
     from pathlib import Path
     from qtpy import QtWidgets
-    from pymodaq_gui.utils.utils import mkQApp
+    from pymodaq_gui.qt_utils import mkQApp
 
     import sys
     app = mkQApp('BeamSteering')
