+++ packages/pymodaq_gui/src/pymodaq_gui/examples/data_picker.py
@@ -3,6 +3,7 @@ from qtpy import QtWidgets
 
 from typing import Optional
 
+import pymodaq_gui.qt_utils
 from pymodaq_data.data import DataRaw, Axis
 
 from pymodaq_gui.utils.widgets.table import SpinBoxDelegate
@@ -94,7 +95,7 @@ class DataPicker(CustomApp):
 
 
 def main():
-    from pymodaq_gui.utils.utils import mkQApp
+    from pymodaq_gui.qt_utils import mkQApp
     import numpy as np
 
     app = mkQApp('DataPicker')
@@ -117,7 +118,7 @@ def main():
 
     data_picker.show_data(data_to_plot)
     win.show()
-    app.exec()
+    pymodaq_gui.qt_utils.exec()
 
 
 if __name__ == '__main__':
