+++ packages/pymodaq_gui/src/pymodaq_gui/examples/parameter_ex.py
@@ -6,6 +6,7 @@ import sys
 from qtpy import QtWidgets, QtCore
 from collections import OrderedDict
 
+import pymodaq_gui.qt_utils
 import pymodaq_gui.utils.widgets.table as table
 from pymodaq_gui.utils.utils import create_nested_menu
 from pymodaq_gui.parameter.pymodaq_ptypes import GroupParameter, registerParameterType
@@ -182,7 +183,7 @@ class ParameterEx(ParameterManager):
 
 
 def main():
-    from pymodaq_gui.utils.utils import mkQApp
+    from pymodaq_gui.qt_utils import mkQApp
 
     app = mkQApp('Parameters')
 
@@ -194,7 +195,7 @@ def main():
     ptree.settings.child('itemss', 'itemsbis').setValue(dict(all_items=['item1', 'item2', 'item3'],
                                                              selected=['item1', 'item3']))
 
-    app.exec()
+    pymodaq_gui.qt_utils.exec()
 
 
 if __name__ == '__main__':
