+++ packages/pymodaq_gui/src/pymodaq_gui/utils/widgets/spinbox.py
@@ -1,6 +1,8 @@
 from qtpy import QtWidgets, QtGui, QtCore
 from pyqtgraph.widgets.SpinBox import SpinBox
 
+import pymodaq_gui.qt_utils
+
 
 class SpinBox(SpinBox):
     """
@@ -48,7 +50,7 @@ class QSpinBox_ro(SpinBox):
 
 if __name__ == '__main__':
 
-    from pymodaq_gui.utils.utils import mkQApp
+    from pymodaq_gui.qt_utils import mkQApp
 
     app = mkQApp('Test Spinbox')
 
@@ -61,4 +63,4 @@ if __name__ == '__main__':
     spinbox.shortcut['Ctrl+Enter'].activatedAmbiguously.connect(lambda: print_spinbox(f'Ctrl+Enter: {spinbox.value()}'))
     spinbox.show()
 
-    app.exec()
+    pymodaq_gui.qt_utils.exec()
